+++ NEWS.md
@@ -30,6 +30,11 @@ Extensively updated documentation for each distribution to include additional de
 * Fixed `dist_negative_binomial()` calculation of skewness and kurtosis for p!=0.5.
 * Fixed floating point issue that caused `dist_mixture()` to incorrectly throw an error when weights did sum to 1 (@statasaurus, #134)
 
+## Breaking changes
+
+* The `type` argument for `dist_multivariate_normal()`'s `quantile()` method 
+  has been renamed to `kind` to avoid name clashes with `quantile.default()`.
+
 # distributional 0.5.0
 
 ## New features